requirements.txt
astroid==2.7.3
black==21.9b0
click==8.0.1
Flask==2.0.1
gunicorn==20.1.0
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.3.0
psycopg2-binary==2.9.1
pylint==2.10.2
python-dotenv==0.19.0
regex==2021.8.28
rope==0.19.0
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
Werkzeug==2.0.1
wrapt==1.12.1
